worker/requirements.txt
redis==5.0.1
Pillow==10.1.0
